Refactor to extract MidiParser class
