+++ src/hid/midi.h
@@ -2,15 +2,12 @@
 #ifndef DSY_MIDI_H
 #define DSY_MIDI_H
 
-// TODO: make this adjustable
-#define SYSEX_BUFFER_LEN 128
-
 #include <stdint.h>
 #include <stdlib.h>
 #include "per/uart.h"
 #include "util/ringbuffer.h"
 #include "util/FIFO.h"
-#include "hid/MidiEvent.h"
+#include "hid/midi_parser.h"
 #include "hid/usb_midi.h"
 #include "sys/dma.h"
 #include "sys/system.h"
@@ -152,12 +149,8 @@ class MidiHandler
     void Init(Config config)
     {
         config_ = config;
-
         transport_.Init(config_.transport_config);
-
-        //event_q_.Init();
-        incoming_message_.type = MessageLast;
-        pstate_                = ParserEmpty;
+        parser_.Init();
     }
 
     /** Starts listening on the selected input mode(s). MidiEvent Queue will begin to fill, and can be checked with HasEvents() */
@@ -174,7 +167,7 @@ class MidiHandler
         // Flush the buff, and restart.
         if(!transport_.RxActive())
         {
-            pstate_ = ParserEmpty;
+            parser_.Reset();
             transport_.FlushRx();
             StartReceive();
         }
@@ -202,204 +195,30 @@ class MidiHandler
     }
 
     /** Feed in bytes to state machine from a queue.
-    Populates internal FIFO queue with MIDI Messages
-    For example with uart:
-    midi.Parse(uart.PopRx());
-    \param byte &
+        Populates internal FIFO queue with MIDI Messages
+        For example with uart:
+            midi.Parse(uart.PopRx());
+        \param byte &
     */
     void Parse(uint8_t byte)
     {
-        // reset parser when status byte is received
-        if((byte & kStatusByteMask) && pstate_ != ParserSysEx)
+        MidiEvent event;
+        if(parser_.Parse(byte, &event))
         {
-            pstate_ = ParserEmpty;
-        }
-        switch(pstate_)
-        {
-            case ParserEmpty:
-                // check byte for valid Status Byte
-                if(byte & kStatusByteMask)
-                {
-                    // Get MessageType, and Channel
-                    incoming_message_.channel = byte & kChannelMask;
-                    incoming_message_.type    = static_cast<MidiMessageType>(
-                        (byte & kMessageMask) >> 4);
-
-                    // Validate, and move on.
-                    if(incoming_message_.type < MessageLast)
-                    {
-                        pstate_ = ParserHasStatus;
-                        // Mark this status byte as running_status
-                        running_status_ = incoming_message_.type;
-
-                        if(running_status_ == SystemCommon)
-                        {
-                            incoming_message_.channel = 0;
-                            //system real time = 1111 1xxx
-                            if(byte & 0x08)
-                            {
-                                incoming_message_.type = SystemRealTime;
-                                running_status_        = SystemRealTime;
-                                incoming_message_.srt_type
-                                    = static_cast<SystemRealTimeType>(
-                                        byte & kSystemRealTimeMask);
-
-                                //short circuit to start
-                                pstate_ = ParserEmpty;
-                                event_q_.PushBack(incoming_message_);
-                            }
-                            //system common
-                            else
-                            {
-                                incoming_message_.sc_type
-                                    = static_cast<SystemCommonType>(byte
-                                                                    & 0x07);
-                                //sysex
-                                if(incoming_message_.sc_type == SystemExclusive)
-                                {
-                                    pstate_ = ParserSysEx;
-                                    incoming_message_.sysex_message_len = 0;
-                                }
-                                //short circuit
-                                else if(incoming_message_.sc_type > SongSelect)
-                                {
-                                    pstate_ = ParserEmpty;
-                                    event_q_.PushBack(incoming_message_);
-                                }
-                            }
-                        }
-                    }
-                    // Else we'll keep waiting for a valid incoming status byte
-                }
-                else
-                {
-                    // Handle as running status
-                    incoming_message_.type    = running_status_;
-                    incoming_message_.data[0] = byte & kDataByteMask;
-                    //check for single byte running status, really this only applies to channel pressure though
-                    if(running_status_ == ChannelPressure
-                       || running_status_ == ProgramChange
-                       || incoming_message_.sc_type == MTCQuarterFrame
-                       || incoming_message_.sc_type == SongSelect)
-                    {
-                        //Send the single byte update
-                        pstate_ = ParserEmpty;
-                        event_q_.PushBack(incoming_message_);
-                    }
-                    else
-                    {
-                        pstate_
-                            = ParserHasData0; //we need to get the 2nd byte yet.
-                    }
-                }
-                break;
-            case ParserHasStatus:
-                if((byte & kStatusByteMask) == 0)
-                {
-                    incoming_message_.data[0] = byte & kDataByteMask;
-                    if(running_status_ == ChannelPressure
-                       || running_status_ == ProgramChange
-                       || incoming_message_.sc_type == MTCQuarterFrame
-                       || incoming_message_.sc_type == SongSelect)
-                    {
-                        //these are just one data byte, so we short circuit back to start
-                        pstate_ = ParserEmpty;
-                        event_q_.PushBack(incoming_message_);
-                    }
-                    else
-                    {
-                        pstate_ = ParserHasData0;
-                    }
-
-                    //ChannelModeMessages (reserved Control Changes)
-                    if(running_status_ == ControlChange
-                       && incoming_message_.data[0] > 119)
-                    {
-                        incoming_message_.type = ChannelMode;
-                        running_status_        = ChannelMode;
-                        incoming_message_.cm_type
-                            = static_cast<ChannelModeType>(
-                                incoming_message_.data[0] - 120);
-                    }
-                }
-                else
-                {
-                    // invalid message go back to start ;p
-                    pstate_ = ParserEmpty;
-                }
-                break;
-            case ParserHasData0:
-                if((byte & kStatusByteMask) == 0)
-                {
-                    incoming_message_.data[1] = byte & kDataByteMask;
-
-                    //velocity 0 NoteOns are NoteOffs
-                    if(running_status_ == NoteOn
-                       && incoming_message_.data[1] == 0)
-                    {
-                        incoming_message_.type = NoteOff;
-                    }
-
-                    // At this point the message is valid, and we can add this MidiEvent to the queue
-                    event_q_.PushBack(incoming_message_);
-                }
-                else
-                {
-                    // invalid message go back to start ;p
-                    pstate_ = ParserEmpty;
-                }
-                // Regardless, of whether the data was valid or not we go back to empty
-                // because either the message is queued for handling or its not.
-                pstate_ = ParserEmpty;
-                break;
-            case ParserSysEx:
-                // end of sysex
-                if(byte == 0xf7)
-                {
-                    pstate_ = ParserEmpty;
-                    event_q_.PushBack(incoming_message_);
-                }
-                else if(incoming_message_.sysex_message_len < SYSEX_BUFFER_LEN)
-                {
-                    incoming_message_
-                        .sysex_data[incoming_message_.sysex_message_len]
-                        = byte;
-                    incoming_message_.sysex_message_len++;
-                }
-                break;
-            default: break;
+            event_q_.PushBack(event);
         }
     }
 
   private:
-    enum ParserState
-    {
-        ParserEmpty,
-        ParserHasStatus,
-        ParserHasData0,
-        ParserSysEx,
-    };
-    ParserState          pstate_;
-    MidiEvent            incoming_message_;
-    FIFO<MidiEvent, 256> event_q_;
-    uint32_t             last_read_; // time of last byte
-    MidiMessageType      running_status_;
     Config               config_;
     Transport            transport_;
-
-    // Masks to check for message type, and byte content
-    const uint8_t kStatusByteMask     = 0x80;
-    const uint8_t kMessageMask        = 0x70;
-    const uint8_t kDataByteMask       = 0x7F;
-    const uint8_t kSystemCommonMask   = 0xF0;
-    const uint8_t kChannelMask        = 0x0F;
-    const uint8_t kRealTimeMask       = 0xF8;
-    const uint8_t kSystemRealTimeMask = 0x07;
+    MidiParser           parser_;
+    FIFO<MidiEvent, 256> event_q_;
 
     static void ParseCallback(uint8_t *data, size_t size, void *context)
     {
-        MidiHandler *handler = reinterpret_cast<MidiHandler*>(context);
-        for (size_t i = 0; i < size; i++)
+        MidiHandler* handler = reinterpret_cast<MidiHandler*>(context);
+        for(size_t i = 0; i < size; i++)
         {
             handler->Parse(data[i]);
         }
